Remove interpolation
Scores in `emoji-emotion` run from 3 (most positive) to -4 (most
negative).  Previously, you saw a score from 5 to -5 (just like in
AFINN), as the build script interpolated those words to that scale.
That however resulted in “holes” (such as `3` never appearing as a
score).  Lastly, it meant that when the most positive score was
removed (`:triumph` in c29abc5), all scores changed.

I think this is bad, plus I’m planning on adding more emoji, and I’d
rather have the scores be absolute and not changing if more emoji
are added.